+++ kaskade/cli.py
@@ -1,24 +1,31 @@
 import sys
+from logging import DEBUG
 
 from confluent_kafka import KafkaException
 from rich.console import Console
 
-from kaskade import APP_UI_LOG, logger
+from kaskade import logger
 from kaskade.config import Config
 from kaskade.emojis import THINKING_FACE
+from kaskade.renderables.config_examples import ConfigExamples
+from kaskade.renderables.kaskade_info import KaskadeInfo
 from kaskade.renderables.kaskade_name import KaskadeName
 from kaskade.renderables.kaskade_version import KaskadeVersion
 from kaskade.tui import Tui
 
 
 class Cli:
-    def __init__(self, print_version: bool, config_file: str) -> None:
+    def __init__(
+        self, print_version: bool, print_information: bool, config_file: str
+    ) -> None:
         self.print_version = print_version
+        self.print_information = print_information
         self.config_file = config_file
 
     def run(self) -> None:
         try:
             self.print_version_option()
+            self.print_information_option()
             self.run_tui()
         except Exception as ex:
             message = str(ex)
@@ -37,13 +44,36 @@ class Cli:
 
             sys.exit(1)
 
+    def print_information_option(self) -> None:
+        if not self.print_information:
+            return
+
+        console = Console()
+        console.print(KaskadeName())
+        console.print()
+        console.print(KaskadeInfo())
+        console.print()
+        console.print(ConfigExamples())
+
+        sys.exit(0)
+
     def print_version_option(self) -> None:
-        if self.print_version:
-            console = Console()
-            console.print(KaskadeName())
-            console.print(KaskadeVersion())
-            sys.exit(0)
+        if not self.print_version:
+            return
+
+        console = Console()
+        console.print(KaskadeName())
+        console.print(KaskadeVersion())
+
+        sys.exit(0)
 
     def run_tui(self) -> None:
         config = Config(self.config_file)
-        Tui.run(config=config, log=APP_UI_LOG if config.kaskade.get("log-ui") else None)
+        is_debug = bool(config.kaskade.get("debug"))
+
+        if is_debug:
+            logger.setLevel(DEBUG)
+            logger.debug("Starting in debug mode")
+
+        logger.debug("Starting TUI")
+        Tui.run(config=config)