+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.0] - 2021-11-19
 ### Added
 - Error message panel
 - Home folder
@@ -30,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Testing release
 - Showing the messages count and lag
 
-[Unreleased]: https://github.com/sauljabin/kaskade/compare/v0.2.9...HEAD
+[Unreleased]: https://github.com/sauljabin/kaskade/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/sauljabin/kaskade/compare/v0.2.9...v0.3.0
 [0.2.9]: https://github.com/sauljabin/kaskade/compare/0.2.8...v0.2.9
 [0.2.8]: https://github.com/sauljabin/kaskade/compare/0.2.7...0.2.8